requirements.txt
python-telegram-bot==21.5
aiosqlite==0.20.0
aiohttp==3.10.5
python-dotenv==1.0.1
